fauctdrop-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.15.3
web3==7.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.44
pydantic==2.9.2
python-multipart==0.0.20
httpx==0.28.1
eth-account==0.13.7
playwright==1.49.1
pytest-playwright==0.6.2
shortuuid==1.0.13
alchemy-sdk==0.1.1
dataclass-wizard==0.22.3
playwright-stealth==1.0.6
Authlib==1.3.1
itsdangerous==2.2.0
asyncpg==0.29.0